+++ ver3.0(choi)/execution/.eslintrc.json
@@ -1,0 +1,31 @@
+{
+  "env": {
+    "node": true,
+    "es6": true
+  },
+  "extends": ["eslint:recommended", "plugin:prettier/recommended", "prettier"],
+  "parserOptions": {
+    "ecmaVersion": "latest",
+    "sourceType": "module"
+  },
+  "plugins": ["prettier"],
+  "rules": {
+    "prettier/prettier": [
+      "error",
+      {
+        "printWidth": 200,
+        "singleQuote": true,
+        "trailingComma": "es5",
+        "tabWidth": 2,
+        "semi": true,
+        "arrowParens": "always",
+        "endOfLine": "lf"
+      }
+    ],
+    "no-unused-vars": "warn"
+  },
+  "ignorePatterns": ["__tests__/**/*", "node_modules/"],
+  "globals": {
+    "BigInt": "readonly"
+  }
+}
