+++ ver3.0(choi)/execution/.prettierrc
@@ -1,0 +1,9 @@
+{
+  "printWidth": 200,
+  "singleQuote": true,
+  "trailingComma": "es5",
+  "tabWidth": 2,
+  "semi": true,
+  "arrowParens": "always",
+  "endOfLine": "lf"
+}
