+++ ver3.0(choi)/execution/eslintrc.yml
@@ -1,0 +1,16 @@
+env:
+  jest: true # Jest 글로벌
+plugins:
+  - jest # Jest 테스트를 위해 플러그인이 필요하다.
+rules:
+  # Jest Eslint 옵션은 0,1,2 (off, warn, error) 만 옵션으로 사용 가능하다.
+  jest/no-disabled-tests:
+    - warn
+  jest/no-focused-tests:
+    - error
+  jest/no-identical-title:
+    - error
+  jest/prefer-to-have-length:
+    - warn
+  jest/valid-expect:
+    - error
