+++ ver3.0(choi)/execution/package.json
@@ -1,0 +1,46 @@
+{
+  "name": "main",
+  "version": "1.0.0",
+  "main": "index.js",
+  "license": "MIT",
+  "type": "module",
+  "scripts": {
+    "dev": "nodemon src/app.js",
+    "lint": "eslint ."
+  },
+  "dependencies": {
+    "@prisma/client": "^5.11.0",
+    "argon2": "^0.40.1",
+    "axios": "^1.6.8",
+    "connect-redis": "^7.1.1",
+    "cookie-parser": "^1.4.6",
+    "cors": "^2.8.5",
+    "express": "^4.19.1",
+    "express-mysql-session": "^3.0.0",
+    "express-session": "^1.18.0",
+    "joi": "^17.12.2",
+    "kafkajs": "^2.2.4",
+    "mysql": "^2.18.1",
+    "node-schedule": "^2.1.1",
+    "nodemailer": "^6.9.13",
+    "passport": "^0.7.0",
+    "passport-google-oauth20": "^2.0.0",
+    "passport-kakao": "^1.0.1",
+    "passport-local": "^1.0.0",
+    "passport-naver-v2": "^2.0.8",
+    "prisma": "^5.11.0",
+    "prom-client": "^15.1.1",
+    "redis": "^4.6.13",
+    "winston": "^3.13.0",
+    "winston-daily-rotate-file": "^5.0.0",
+    "ws": "^8.16.0"
+  },
+  "devDependencies": {
+    "dotenv": "^16.4.5",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "nodemon": "^3.1.0",
+    "prettier": "^3.2.5"
+  }
+}
