+++ ver3.0(choi)/execution/src/middlewares/checkLoginMiddleware.js
@@ -1,0 +1,7 @@
+export default function isNotLogin(req, res, next) {
+  if (!req.isAuthenticated()) {
+    next();
+  } else {
+    return res.json({ message: '이미 로그인 된상태 입니다' });
+  }
+}
