+++ ver3.0(choi)/execution/src/middlewares/generalError.middleware.js
@@ -1,0 +1,10 @@
+import { logger } from './log.middleware.js';
+
+export default (err, req, res, next) => {
+  logger.error(`Error Status: ${err.status}, Message: ${err.message}`);
+  return res.status(err.status || 500).json({
+    error: {
+      message: err.message,
+    },
+  });
+};
