+++ ver3.0(choi)/execution/src/middlewares/notFoundError.middleware.js
@@ -1,0 +1,5 @@
+export default (req, res, next) => {
+  const error = new Error(`${req.method} ${req.url} 라우터가 없습니다.`);
+  error.status = 404;
+  next(error);
+};
