+++ ver3.0(choi)/execution/src/utils/prisma/index.js
@@ -1,0 +1,9 @@
+import { PrismaClient } from '@prisma/client';
+
+export const prisma = new PrismaClient({
+  // Prisma를 이용해 데이터베이스를 접근할 때, SQL을 출력해줍니다.
+  log: [],
+
+  // 에러 메시지를 평문이 아닌, 개발자가 읽기 쉬운 형태로 출력해줍니다.
+  errorFormat: 'pretty',
+}); // PrismaClient 인스턴스를 생성합니다.
