+++ ver3.0(choi)/main/src/app.js
@@ -17,12 +17,15 @@ import router from './routes/index.js';
 import { gameTotal } from './utils/schedule/gameTotal.js';
 import { gameSetting } from './utils/schedule/gameSetting.js';
 import { createDummyEvent } from './utils/schedule/gameMiddle.js';
+import { initializeProducer, terminateProducer } from './utils/kafkaProducer/kafkaProducer.js';
 
 dotenv.config();
 
 const app = express();
 const server = createServer(app);
 const PORT = process.env.PORT;
+
+initializeProducer().catch(console.error);
 
 //app.use(LogMiddleware);
 app.use(
@@ -112,3 +115,14 @@ app.use(generalErrorHandler);
 server.listen(PORT, () => {
   console.log(PORT, '포트로 서버가 열렸어요!');
 });
+
+// 애플리케이션 종료 시 Kafka Producer 연결 해제
+const gracefulShutdown = async () => {
+  console.log('Shutting down gracefully...');
+  await terminateProducer();
+  console.log('Kafka Producer disconnected');
+  process.exit(0);
+};
+
+process.on('SIGINT', gracefulShutdown);
+process.on('SIGTERM', gracefulShutdown);
