+++ ver3.0(choi)/main/src/routes/order/order.controller.js
@@ -1,4 +1,4 @@
-import { insertOrderMessageQueue } from '../../utils/orderQueue/index.js';
+import { sendMessage } from '../../utils/kafkaProducer/kafkaProducer.js';
 export class OrderController {
   constructor(orderService) {
     this.orderService = orderService;
@@ -80,7 +80,9 @@ export class OrderController {
         price: orderData.price,
       };
       const jsonOrderDataString = JSON.stringify(jsonOrderData);
-      insertOrderMessageQueue(jsonOrderDataString);
+      console.log('주문이 접수되었습니다.');
+      await sendMessage('executionQueue', [{ value: jsonOrderDataString }]);
+      console.log('주문을 kafka서버로 전송합니다.');
       return res.json({ message: '주문이 접수 되었습니다.' });
     } catch (error) {
       console.log(error.stack);
@@ -144,7 +146,7 @@ export class OrderController {
         price: correctedPrice,
       };
       const jsonOrderDataString = JSON.stringify(jsonOrderData);
-      insertOrderMessageQueue(jsonOrderDataString);
+      await sendMessage('executionQueue', [{ value: jsonOrderDataString }]);
       return res.json({ message: '주문이  접수되었습니다.' });
     } catch (error) {
       console.log(error.stack);
